requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
openai>=1.0.0
sqlalchemy>=2.0.0
chromadb>=0.5.0
faker>=25.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
langfuse>=2.0.0
pypdf>=4.0.0
python-docx>=1.0.0
pandas>=2.0.0
ipywidgets>=8.0.0
matplotlib>=3.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
